+++ client/bindings/nodejs/types/outputBuilderOptions/nftOutputOptions.ts
@@ -1,3 +1,4 @@
+import type { HexEncodedString } from '@iota/iota.js';
 import type { FeatureTypes } from '@iota/types';
 import type { IBasicOutputBuilderOptions } from './basicOutputOptions';
 
@@ -5,6 +6,6 @@ import type { IBasicOutputBuilderOptions } from './basicOutputOptions';
  * Options for building an Nft Output
  */
 export interface INftOutputBuilderOptions extends IBasicOutputBuilderOptions {
-    nftId: string;
+    nftId: HexEncodedString;
     immutableFeatures?: FeatureTypes[];
 }